railway/requirements.txt
httpx>=0.27.0
python-dotenv>=1.0.0
supabase>=2.0.0
playwright>=1.40.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
schedule>=1.2.2
flask>=3.0.0
gunicorn>=21.0.0
